+++ src/AdjustingDetectors.h
@@ -11,7 +11,7 @@ class AdjustingDetectors : public AutoLdrSpotDetectors<AdjustingLDR>
   int thresholdLevel;
 
 public:
-  AdjustingDetectors(SensorChangeAction & action, const std::initializer_list<int> & il, int thresholdLevel = 100);
+  AdjustingDetectors(SensorChangeAction & action, const std::initializer_list<uint8_t> & il, int thresholdLevel = 100);
 
   int getThresholdLevel() { return thresholdLevel; }
