+++ src/ThresholdDetectors.cpp
@@ -1,7 +1,7 @@
 #include "ThresholdDetectors.h"
 #include "ThresholdLDR.h"
 
-ThresholdDetectors::ThresholdDetectors(SensorChangeAction & action, const std::initializer_list<int> & il, int threshold)
+ThresholdDetectors::ThresholdDetectors(SensorChangeAction & action, const std::initializer_list<uint8_t> & il, int threshold)
   : AutoLdrSpotDetectors(action, il)
 {
   for (unsigned int i = 0 ; i < ldrCount ; ++i)
