+++ src/ThresholdDetectors.h
@@ -6,7 +6,7 @@
 class ThresholdDetectors : public AutoLdrSpotDetectors<ThresholdLDR>
 {
 public:
-  ThresholdDetectors(SensorChangeAction & action, const std::initializer_list<int> & il, int threshold);
+  ThresholdDetectors(SensorChangeAction & action, const std::initializer_list<uint8_t> & il, int threshold);
 
   virtual void update() override;
 };
